Move FFT logic to fft.c, fft.h.
